edesign.OD(maSigPro)	R Documentation

Experimental design with a measured independent variable

Description

code{edesign.OD} contains the experimental design of a E.coli growth time course microarray experiment with a temperature shift treatment. The OD of each culture was measured and used in the experimental design as independent variable.

Usage

data(edesign.OD)

Format

A data frame with 52 rows and the following 4 variables.

OD

a numeric vector. Indicates the OD value of the sampled culture

Replicate

a numeric vector

37

a numeric vector. No temperature shitf treatment

SHIFT

a numeric vector. Temperature shift treatment

Examples


data(edesign.OD)
## maybe str(edesign.OD) ; plot(edesign.OD) ...

[Package maSigPro version 1.0 Index]